requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
more-itertools==4.2.0
multidict==4.3.1
pluggy==0.6.0
py==1.5.4
pytest==3.6.3
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
vcrpy==1.13.0
wrapt==1.10.11
yarl==1.2.6
